+++ docs/lib/menus.js
@@ -1,3 +1,10 @@
+export const mainNavigation = [
+   { name: 'Overview', href: '/' },
+   { name: 'Technical Documentation', href: '/data-access' },
+   { name: 'Other Information', href: '/nccid-collaborative' },
+   { name: 'Experiments', href: '/experiments' },
+]
+
 export const techNavigation = [
    { name: 'Accessing the NCCID data', href: '/data-access' },
    { name: 'Resources for data users', href: '/resources' },
@@ -9,13 +16,6 @@ export const otherNavigation = [
    { name: 'Performance Assessment Call', href: '/performance-assessment-call' },
    { name: 'Links to other datasets', href: '/other-datasets' },
    { name: 'Training data statistics', href: '/stats' },
-]
-
-export const mainNavigation = [
-   { name: 'Overview', href: '/' },
-   { name: 'Technical Documentation', href: '/data-access' },
-   { name: 'Other Information', href: '/nccid-collaborative' },
-   { name: 'Statistics', href: '/statistics' },
 ]
 
 export const footerNavigation = [
